requirements.txt
flet
bcrypt
PyJWT
pytest
mysql-connector-python
python-dotenv